All Consecutive Elements in Java
Given an unsorted array of numbers, write a function that returns true if complete array consists of consecutive numbers. That means if we sort the array, all the elements are consecutive. Array contains only positive numbers.
Try doing it in O(n), where n is the size of given array.